move markdown file to plugins
